refactor(mode): move tmux render pipeline to src/render
