+++ src/input/tmux/mouse.rs
@@ -17,8 +17,8 @@ use crate::app_mode::TmuxActiveSet;
 use crate::configs::OzmuxConfigsResource;
 use crate::input::InputPhase;
 use crate::input::gesture::ClickTracker;
-use crate::mode::tmux::copy_mode::{CopyModeSnapshot, cell_at_pane};
-use crate::mode::tmux::render::{DividerPixelRect, PackedTmuxLayout};
+use crate::render::tmux::copy_mode::{CopyModeSnapshot, cell_at_pane};
+use crate::render::tmux::{DividerPixelRect, PackedTmuxLayout};
 use crate::ui::copy_mode::CopyModeState;
 use crate::ui::copy_search::CopyPrompt;
 use bevy::ecs::system::SystemParam;
