+++ src/render.rs
@@ -1,0 +1,4 @@
+//! Terminal-grid rendering features. tmux-only today: the pane layout → grid
+//! pipeline, paint rescue, and capture-driven copy-mode rendering.
+
+pub(crate) mod tmux;
